requirements.txt
amqp==2.3.2
appdirs==1.4.3
asgiref==3.4.1
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
bcrypt==3.1.6
billiard==3.5.0.4
black==20.8b1
blinker==1.4
bugsnag==3.6.0
cached-property==1.5.1
celery==4.2.1
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
click==7.1.2
colorama==0.3.7
configobj==5.0.6
constantly==15.1.0
convertdate==2.3.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
defusedxml==0.6.0
Django==2.1.2
django-base64field==1.0
django-cors-headers==2.4.0
django-dotenv==1.4.2
django-filter==2.0.0
django-next-prev==1.0.1
django-redis==4.10.0
django-rest-passwordreset==1.0.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==3.3
elastic-transport==8.1.2
elasticsearch==8.2.2
et-xmlfile==1.0.1
ez-setup==0.9
h11==0.8.1
h2==3.2.0
hijri-converter==2.1.1
holidays==0.10.5.2
hpack==3.0.0
http3==0.6.7
httplib2==0.9.2
hyperframe==5.2.0
hyperlink==17.3.1
idna==2.6
incremental==16.10.1
isodate==0.6.0
itypes==1.1.0
jdcal==1.4
jeepney==0.4.1
Jinja2==2.7.3
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.4.0
keyring==10.6.0
keyrings.alt==3.0
kombu==4.2.1
korean-lunar-calendar==0.2.1
lxml==4.4.2
MarkupSafe==1.0
mypy-extensions==0.4.3
mysqlclient==2.0.1
ntlm-auth==1.4.0
numpy==1.21.5
oauthlib==2.0.6
openpyxl==3.0.7
pandas==1.1.5
paramiko==2.4.2
pathspec==0.7.0
Pillow==6.0.0
pyasn1==0.4.5
pycparser==2.19
pydash==4.7.5
PyJWT==1.7.1
PyMeeus==0.5.0
PyMySQL==0.9.2
PyNaCl==1.3.0
pyodbc==4.0.30
pyPdf==1.13
PyPDF2==1.26.0
pysftp==0.2.9
python-dateutil==2.8.1
pytz==2018.7
redis==3.0.1
regex==2020.2.20
reportlab==3.5.23
requests==2.20.1
requests-async==0.6.2
requests-ntlm==1.1.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
SecretStorage==3.1.1
six==1.11.0
tk==0.1.0
toml==0.10.1
twilio==6.51.0
typed-ast==1.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.0
urllib3==1.26.9
vine==1.1.4
WebOb==1.8.6
XlsxWriter==1.1.2
zeep==3.4.0
zplgrf==1.6.0
img2pdf
extract-msg==0.29.1
treepoem